Add support for updating packages
This adds the support for updating packages installed using pipsi. There
is currently an assumption that pipsi > 0.5 is being used, as this is
using the `pipsi update` command.

Packages are updated when possible when the state is specified as
'latest'. Where are package is not installed, the regular install will
take place, while already installed packages will attempt an upgrade.